+++ example/lib/main.dart
@@ -1,13 +1,10 @@
 import 'package:flutter/material.dart';
 
-import 'screens/ex_expandable_tree_expanded_screen.dart';
-import 'screens/ex_expandable_tree_screen.dart';
-import 'screens/ex_lazy_stack_screen.dart';
-import 'screens/ex_lazy_tree_single_choice.dart';
-import 'screens/ex_stack_screen.dart';
-import 'screens/ex_tree_single_choice.dart';
-import 'screens/ex_tree_vietnam_regions.dart';
-import 'screens/vts/ex_vts_department_tree_screen.dart';
+import 'screens/expandable_multiple_choice/expandable_multiple_choice_screen.dart';
+import 'screens/expandable_single_choice/expandable_single_choice_screen.dart';
+import 'screens/stack/stack_screen.dart';
+import 'screens/vts/vts_screen.dart';
+import 'widgets/custom_outlined_btn.dart';
 
 void main() {
   runApp(const MainApp());
@@ -31,85 +28,50 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(title: const Text("Tree View Flutter Example")),
-      body: Center(
+      body: const Center(
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: [
-            _buildButton(
-              context,
-              const ExStackScreen(),
-              "Stack Tree\nmultiple choice - parse data 1 time",
-            ),
-            _buildButton(
-              context,
-              const ExLazyStackScreen(),
-              "Lazy Stack Tree\nmultiple choice - parse data run-time",
+            CustomOutlinedButton(
+              screen: StackScreen(),
+              title: "Stack",
             ),
             //_____________________
-            const Divider(
+            Divider(
               thickness: 2,
               height: 50,
               color: Colors.black,
             ),
             //_____________________
-            _buildButton(
-              context,
-              const ExExpandableTreeScreen(),
-              "Expandable Tree\nmultiple choice - parse data 1 time",
-            ),
-            _buildButton(
-              context,
-              const ExExpandableTreeExpandedScreen(),
-              "Expandable Tree expanded first\nmultiple choice - parse data 1 time",
-            ),
-            _buildButton(
-              context,
-              const ExVTSDeptTreeScreen(),
-              "VTS Department Tree Screen",
+            CustomOutlinedButton(
+              screen: ExpandableMultipleChoiceScreen(),
+              title: "Expandable Multiple Choice",
             ),
             //_____________________
-            const Divider(
+            Divider(
               thickness: 2,
               height: 50,
               color: Colors.black,
             ),
             //_____________________
-            _buildButton(
-              context,
-              const ExTreeSingleChoice(),
-              "Expandable Tree\nsingle choice - parse data 1 time",
-            ),
-            _buildButton(
-              context,
-              const ExLazyTreeSingleChoice(),
-              "Expandable Tree\nsingle choice - lazy loading",
+            CustomOutlinedButton(
+              screen: ExpandableSingleChoiceScreen(),
+              title: "Expandable Single Choice",
             ),
             //_____________________
-            _buildButton(
-              context,
-              const ExTreeSingleChoice(),
-              "Expandable Tree\nsingle choice - parse data 1 time",
+            Divider(
+              thickness: 2,
+              height: 50,
+              color: Colors.black,
             ),
-            _buildButton(
-              context,
-              const ExTreeVNRegions(),
-              "Vietnam regions tree",
+            //_____________________
+            CustomOutlinedButton(
+              screen: VTSScreen(),
+              title: "VTS Department Tree",
             ),
           ],
         ),
       ),
     );
   }
-
-  Widget _buildButton(
-          BuildContext context, StatefulWidget screen, String title) =>
-      OutlinedButton(
-        onPressed: () => Navigator.of(context).push(
-          MaterialPageRoute(builder: (_) => screen),
-        ),
-        child: Text(
-          title,
-          textAlign: TextAlign.center,
-        ),
-      );
 }